+++ backend/canisters/community/impl/src/updates/edit_message.rs
@@ -5,7 +5,6 @@ use chat_events::EditMessageArgs;
 use community_canister::edit_message::*;
 use oc_error_codes::OCErrorCode;
 use types::{Achievement, EventIndex, OCResult};
-use utils::migrated_user_ids::MigratedUserIds;
 
 #[update(msgpack = true)]
 #[trace]
@@ -36,7 +35,7 @@ fn edit_message_impl(args: Args, state: &mut RuntimeState) -> OCResult {
             finalise_bot_message: false,
             now,
         },
-        &MigratedUserIds::default(),
+        &state.data.migrated_user_ids,
         Some(CommunityEventPusher {
             now,
             rng: state.env.rng(),