+++ backend/canisters/group/impl/src/updates/cancel_p2p_swap.rs
@@ -4,7 +4,6 @@ use canister_api_macros::update;
 use canister_tracing_macros::trace;
 use group_canister::cancel_p2p_swap::*;
 use types::OCResult;
-use utils::migrated_user_ids::MigratedUserIds;
 
 #[update(msgpack = true)]
 #[trace]
@@ -30,7 +29,7 @@ fn cancel_p2p_swap_impl(args: Args, state: &mut RuntimeState) -> OCResult<u32> {
             args.thread_root_message_index,
             args.message_id,
             now,
-            &MigratedUserIds::default(),
+            &state.data.migrated_user_ids,
         )
         .map(|result| {
             state.push_bot_notification(result.bot_notification);